requirements.txt
streamlit
mysql-connector-python
qrcode[pil]
pillow
matplotlib
